requirements.txt
appdirs==1.4.3
backports.functools-lru-cache==1.4
backports.shutil-get-terminal-size==1.0.0
click==6.7
cycler==0.10.0
decorator==4.1.2
enum34==1.1.6
h5py==2.7.1
ipython==5.5.0
ipython-genutils==0.2.0
Keras==1.2.2
Mako==1.0.7
MarkupSafe==1.0
matplotlib==2.1.0
networkx==2.0
numpy==1.13.3
olefile==0.44
opencv-python==3.3.0.10
pathlib2==2.3.0
pexpect==4.3.0
pickleshare==0.7.4
Pillow==4.3.0
prompt-toolkit==1.0.15
ptyprocess==0.5.2
py==1.5.2
pycuda==2017.1.1
Pygments==2.2.0
pyparsing==2.2.0
pytest==3.2.5
python-dateutil==2.6.1
pytools==2017.6
pytz==2017.3
PyWavelets==0.5.2
PyYAML==3.12
scandir==1.6
scikit-cuda==0.5.1
scikit-image==0.13.1
scikit-learn==0.19.1
scipy==1.0.0
simplegeneric==0.8.1
six==1.11.0
sklearn==0.0
subprocess32==3.2.7
Theano==0.8.2
tqdm==4.19.4
traitlets==4.3.2
vlfeat-ctypes==0.1.5
wcwidth==0.1.7
